requirements.txt
Flask==3.0.3
flask_cors==5.0.0
flask_sqlalchemy==3.1.1
openpyxl==3.1.5
PyJWT==2.9.0
redis==6.1.1
Werkzeug==3.0.6
bcrypt==4.0.1
python-dotenv==1.0.0

